\name{powerGWASinteraction}
\alias{powerGWASinteraction}
\title{Depreciated}
\description{This function is depreciated and has been replaced by \code{powerGG} and \code{powerGE}}
\usage{powerGWASinteraction()}
\value{An error message is printed}
\author{Charles Kooperberg, \email{clk@fredhutch.org}}
\examples{ powerGWASinteraction()}
\seealso{powerGG,powerGE}
\keyword{methods}

